Make JRubyJellyScript#createTagScript thread-safe
It's my understanding that TagLibrary#createTagScript() is not
thread-safe. JellyScript should be thread-safe for run(), but it's OK to
be thread-unsafe for parse(). Is it true?

JRubyJellyScript#createTagScript() which calls
TagLibrary#createTagScript() is invoked from other Jelly scripts which
could run run under multi-threaded environment if I understand WebApp
lifecycle correctly. So I add synchronized to
JRubyJellyScript#createTagScript().

It could be wrong since JellyBuilder#createTagScript() in stapler-groovy
does not have synchronize. What's the difference?

Making JRubyJellyScript and JRubyFacet to be thread-safe is important
for maintaining ScriptingContainer instance lifecycle. If it's
thread-safe, singleton instance is enough. If it's not, we need to have
ScriptingContainer pool.