Reducing this down to a warning.
I noticed that cloudbees wasted minute tracker causes this error, and it's a Widget, so if this is an exception, it'll break all the installations as soon as it upgrades.

Since this is meant to be caused by a programmer while in development, it's much better to issue a warning so that users will not be penalized.