Allow Stapler to accept 'kind' in addition to 'stapler-class'.
Today, Jenkins uses a mix of stapler's data-binding and its own.  The former uses a lot of 'stapler-class[-bag]' annotations, whereas the latter uses 'kind'.  Various jelly elements emit both because they aren't sure which is needed.

This change allows Stapler to accept 'kind' in addition to 'stapler-class', so that a single-consistent term may be used for all contexts.

NOTE: The bias here is towards 'kind' because this keyword choice directly affects the grammar of the yaml-project plugin, and is a more user friendly keyword for the YAML DSL.