Using setHeader("Content-Encoding","gzip") as a trigger turns out problemaic
The code was originally using the act of setHeader("Content-Encoding","gzip") as a clue to transparently gzip-compress OutputStream.

Turns out this is problematic when we forward requests to other Servlets, which might come with its own gzip compression logic. JGit SmartHttpOutputStream was a good example of this. These implementations are correct on their own right because servlet spec doesn't make this a responsibility of the container.

In this change, CompressionFilter.activate() is used as a mechanism to trigger compression.