Fix for #141
At the end of LargeText.writeLogTo() the size size is returned back as
a long, but the calculation is done with int math, not long.  So we get
a problem for large files.

This fix enables files longer than Integer.MAX_VALUE to be handled.

Unfortunately the only way to test this is by writing out a file that is
larger than Integer.MAX_VALUE.  This test takes 24 seconds to run on my
MacBookPro 15" (2015), which has a very fast SSD.  To test in any other
way would mean much larger changes to LargeFile, just for test purposes.