Fix skipping for File and Gzip sources of LargeText
File: Seeking past the end is allowed in a RandomAccessFile. The
behavior is desired there for creating files with "holes".

Gzip: When reaching the EOF, skip will return -1. The while loop will
increment the start at every turn. It might eventually overflow and
finish. GZIPInputStream.skip already skips in a loop and we can drop
the outer loop in GzipAwareSession.